Allow `#display_name` of Payload Objects to Fail (#1)
Currently the PayloadObject#display_name method is considered in-fallible

If it does fail, the Delayed Worker will raise an exception and be restarted. Importantly it will NOT
mark the jobs as failing. This leads to a scenario where when the worker boots the same thing will happen.
A worker will pick up this 'poisoned' job since it will be the 'next up' in the queue. It will then raise an exception and restart, causing the loop to continue.

Removing the offending jobs from the queue is the only way to stop this cycle.

This PR fixes that by not assuming display_name is infallible. Since it is configurable by consumers, we shouldn't expect it to always succeed.

Here we take the same approach for any error as was specifically happening for DeserializationErrors. This means we will fall back to trying to read the display_name by parsing the serialized yml. We also add a log message for this case, so it doesn't get completely swallowed

This makes it so errors in #display_name don't cause the workers to crash

* Failing spec for this case where display_name might raise

* Add fix and more specs around the logging behavior

* Give the error message a bit more details